ml/coding-question/requirements.txt
flask
groq
python-dotenv
gunicorn==21.2.0